requirements.txt
aiohttp==3.10.11
langchain>=0.1.14
openai  # or any model provider like openrouter
python-dotenv
